requirements.txt
streamlit
pandas
gspread
oauth2client
google-auth
